User confirmation
Added code so a the user has to confirm everything entered, so that a
mistype doesn't break the entire game.  It needs to be revamped-- it's
kind of cloogy and when the user declines to confirm the input it asks
for too much to be re-entered, but it doesn't have bugs and it's good
enough to actually use if playing a game.